Rework path
